requirements.txt
flask
google-generativeai
gunicorn
numpy